Rename WeekdayAndTime, to be renamed back
