+++ src/filters/weekdayAndTime2/weekdayAndTime2.constants.test.ts
@@ -1,4 +1,4 @@
-import WEEKDAY_AND_TIME from './weekdayAndTime.constants';
+import WEEKDAY_AND_TIME from './weekdayAndTime2.constants';
 
 describe('WEEKDAY_AND_TIME', () => {
     describe('ANY', () => {
