+++ src/filters/weekdayAndTime2/weekdayAndTime2.helpers.ts
