+++ src/filters/weekdayAndTime2/weekdayAndTime2.ts
@@ -1,6 +1,6 @@
 import Filter from 'filters/filter';
-import WEEKDAY_AND_TIME from './weekdayAndTime.constants';
-import { weekdayAndTimeToDateObj } from './weekdayAndTime.helpers';
+import WEEKDAY_AND_TIME from './weekdayAndTime2.constants';
+import { weekdayAndTimeToDateObj } from './weekdayAndTime2.helpers';
 
 export default class WeekdayAndTime extends Filter {
     constructor() {
